+++ Trunk/LRT12/Components/TBridgePD.cpp
@@ -24,21 +24,21 @@ void TBridgePD::Configure()
 	Config * config = Config::GetInstance();
 
 	/* TODO: default values... */
-	m_minPosition 	= config->Get<float>(m_configSection, "minPosition", 0.0f);
-	m_maxPosition	= config->Get<float>(m_configSection, "maxPosition", 0.0f);
+	m_minPosition 	= config->Get<double>(m_configSection, "minPosition", 0.0f);
+	m_maxPosition	= config->Get<double>(m_configSection, "maxPosition", 0.0f);
 	
-	m_maxPowerUp 	= config->Get<float>(m_configSection, "maxPowerUp", 0.0f);
-	m_powerDown		= config->Get<float>(m_configSection, "powerDown", 0.0f);
+	m_maxPowerUp 	= config->Get<double>(m_configSection, "maxPowerUp", 0.0f);
+	m_powerDown		= config->Get<double>(m_configSection, "powerDown", 0.0f);
 	
-	m_pGainUp		= config->Get<float>(m_configSection, "pGainUp", 0.0f);
-	m_pGainDown		= config->Get<float>(m_configSection, "pGainDown", 0.0f);
+	m_pGainUp		= config->Get<double>(m_configSection, "pGainUp", 0.0f);
+	m_pGainDown		= config->Get<double>(m_configSection, "pGainDown", 0.0f);
 	
-	m_powerRetainUp	= config->Get<float>(m_configSection, "powerRetainUp", 0.0f);
+	m_powerRetainUp	= config->Get<double>(m_configSection, "powerRetainUp", 0.0f);
 }
 
 void TBridgePD::Output()
 {
-	float potValue = m_bpdPot->GetAverageValue();
+	double potValue = m_bpdPot->GetAverageValue();
 
 #ifdef USE_DASHBOARD
 	SmartDashboard:Log(potValue, "BPD Pot Value")
@@ -68,9 +68,9 @@ void TBridgePD::Output()
 			}
 			else
 			{
-				float error = m_maxPosition - potValue;
+				double error = m_maxPosition - potValue;
 				
-				m_bpdEsc->SetDutyCycle(Util::Max<float>(Util::Min<float>(m_maxPowerUp*1.5, error*m_pGainUp), 0.15));
+				m_bpdEsc->SetDutyCycle(Util::Max<double>(Util::Min<double>(m_maxPowerUp*1.5, error*m_pGainUp), 0.15));
 			}
 			break;
 		case ACTION::BPD::PRESET_BOTTOM:
@@ -86,9 +86,9 @@ void TBridgePD::Output()
 			{
 				action->bridgePD->completion_status = ACTION::IN_PROGRESS;
 				
-				float error = m_minPosition - potValue;
+				double error = m_minPosition - potValue;
 			
-				m_bpdEsc->SetDutyCycle(Util::Max<float>(m_powerDown, error*m_pGainDown));
+				m_bpdEsc->SetDutyCycle(Util::Max<double>(m_powerDown, error*m_pGainDown));
 			}
 			break;
 	}
